// frontend/src/lib/api/generated/model/departmentDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ระบบรายงานค่าสาธารณูปโภค กรมประมง API
 * OpenAPI spec version: 2.0.0
 */

export interface DepartmentDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  costCenterCode?: string;
  createdAt: string;
  depositUnit?: string;
  disbursingUnit?: string;
  division?: string;
  email?: string;
  fullName: string;
  id: string;
  location?: string;
  phone?: string;
  province?: string;
  responsiblePerson?: string;
  responsiblePhone?: string;
  shortName?: string;
  type?: string;
  updatedAt: string;
}